+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTVC27ZDB8SC42GV29SW0AH2"
+        "01KTX87DJV779XXVR1V6HJYDP2"
     )
     r"""The unique identifier of the data source"""