+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KTVC27ZFWX9ZNBDSDXYJ0ZQ4"
+        "01KTX87DJW1Z99TQCAJJ9HJE6F"
     )
     r"""The unique identifier of the data source"""