+++ packages/orq-rc/RELEASES.md
@@ -8649,3 +8649,13 @@ Based on:
 - [python v4.11.0-rc.14] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.14] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.14 - packages/orq-rc
+
+## 2026-06-12 06:29:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.777.1 (2.903.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.15] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.15] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.15 - packages/orq-rc